requirements.txt
albumentations==1.3.1
matplotlib==3.7.2
matplotlib==3.1.2
numpy==1.24.3
numpy==1.17.4
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.74
Pillow==10.0.0
Pillow==7.0.0
Pillow==10.1.0
tqdm==4.65.0
ultralytics==8.0.147
